Adjust audio balance between MCU audio and Bluetooth audio
Changed DAC config to use FVR at 2.048V as it's positive reference.

This allows the SINE table for tone output to use the full range of 0-255 without nearing limits of the Op Amp buffer.

Circuit was changed to to use parallel unity and inverting unity amplifiers for the differential amplifier for audio output. Previously it was a 1/2 gain feeding into an inverting unity. This simplifies and accounts for the DAC output being about half peak-peak of what it used to be.

The BM62 config was also updated to reduce voice output gain to account for the 1/2 gain amplifier being changed to unity gain.