Improve EEPEOM code
Fix a bug in EEPROM_IsDoneWriting() -  it was not taking into consideration the current deferred write that was being written (resulting in the phone rebooting before EEPROM erase was complete when performing a factory reset).

Add a few while(NVMCON0bits.GO) loops at the beginnings of functions that perform EEPROM read/writes to absolutely guarantee no interference with pending async writes.

Clean up a few unnecessary while(NVMCON0bits.GO) loops that were previously added in hopes of fixing some write bugs (but it was actually incomplete writes due to the EEPROM_IsDoneWriting bug).

Removed the repeatingByte field from the buffered write header to reduce the size of headers (it was a wasted byte for normal non-repeating buffered writes). The repeating byte value is now written to the buffer after the header.

Add checks for zero-length reads/writes to avoid bugs.