Fix bug with EEPROM storage layout/initialization
The "reserved" portion of storage was intended to be directly before the directory and credit card number storage, and initialization code was written based on this assumption. however, I had misplaced the "reserved" field in the struct, which caused a chunk of storage to not be written to EEPROM during initialization.