Switch voice command button from SEND to END
Voice command is now triggered by holding the END button for a short time instead of holding the SEND button.

This allows voice command to be triggered regardless of whether a number is currently typed on the screen, or any other situation where pressing the SEND button will place a call.