Synchronize power-off beep with original transceiver beep.
The timing between the PWR button press/hold and the original transceiver deciding to power off (and starting a beep) is too inconsistent to synchronized purely based on a time delay assumption. The power-off detection and beep behavior has been moved to the TRANSCEIVER module, which now watches for the transceiver to enable the speaker while the PWR button is pressed to detect power-down initiation, then watch for the transceiver to disable the speaker when powering off to know when to stop the power-off beep. Now the phone is guaranteed to power off if you hear the beep then release the PWR button.