Fix bug with Call Waiting and Hands-Free Controller
After answering or ignoring an incoming call waiting (while already in an active call), the Hands-Free Controller incorrectly believed that it was no longer "in a call". The command for enabling the "IN USE" indicator needs to be explicitly sent to the handset again so that the Hands-Free Controller detects it.