Change "no Bluetooth connection" indicator to flashing signal strength bars
The previous behavior of flashing the "NO SVC" indicator for "no Bluetooth connection" was incompatible with the OEM Hands-Free Controller, because the controller believes the phone has service and is ready to make calls whenever the "NO SVC" indicator is off. This would allow for the hands-free controls to be used (if timed right) to attempt a call while there is no Bluetooth connection.

Now the "NO SVC" indicator is on solid and all signal strength bars are flashing when there is no Bluetooth connection.