Update circuit layout, README, and datasheets.
Update circuit layout, README, and datasheets to adjust for the components I ended up ordering, and adjusting the layout to fid in the project box I ordered. The biggest change is trading out the pair of INA105 op amps for single DRV134 differential audio driver. I'm also switching to a different (but similar) transistors whose pinout is reversed.