Fix sound interruption when ending a call
Fix a button beep sound interruption when ending a call caused by a specific sequence of commands that are needed for OEM Hands-Free Controller compatibility (to deal with switching between car audio output and phone audio output).

When NOT using the OEM Hands-Free Controller, these additional commands are unnecessary and interrupt sound output.

Additional commands are now executed only when the OEM Hands-Free Controller is believed to be connected.