Update pairing to use 4-digit PIN
When pairing the Bluetooth to a cell phone, a 4-digit PIN is now displayed on the car phone handset that must be entered on the cell phone that it is pairing to.

Entering pairing mode on the car phone now erases memory of the currently paired cell phone.

Removed the PIN-protected ability to erase memory of the currently paired cell phone when viewing the paired cell phone Bluetooth name (simply enter and exit pairing mode if you want to clear this out).

Also changed format of BM62 configuration to be a "UI Patch" instead of a "Full EEPROM" to avoid overwriting the BM62's unique Bluetooth Address (BD_ADDR) with a default address.