Another attempt to fix strange mic audio issues caused by disabling AEC
The BM62 doesn't seem to like handling commands to enable/disable AEC too soon after entering a state where the microphone is used (SCO connected and in a call or in voice command).

A delay of 1 second seems to make it reliable. Afterward, no delay is needed when switching between handset and hands-free mic. But more testing is needed.