Keep handset powered on during restarts
Inverted the switched power enable signal from the MCU with a transistor so that switched power is on by default unless the MCU explicitly disables it. This allows everything to remain powered on while the MCU reboots for a restart. If the MCU is rebooting to power off, then it immediately turns switched power off before entering the sleep state.

This is simpler and more robust than attempting to write code to fully reinitialize ALL MCU state/memory except for switched power enable output.