Simplify load switch enable circuit, and add 100uF cap to 12V supply
Eliminated the load switch enable inverting transistor and simplified to a combination of a pull-up resistor on the enable pin with open-drain and initially high configuration of the MCU pin.

This still keeps the load switches enabled during a MCU reboot like the previous design, but with fewer components and less wasted power.

Also added a 100uF capacitor to the 12V supply to help reduce voltage dips.